Move `FeedLoaderPresentationAdapter` to separate file
